+++ scripts/clear_results.sh
@@ -1,21 +1,31 @@
 #!/usr/bin/env bash
 
-## DEPENDENCY
-pandora_helper="/mnt/archgen/tools/helper_scripts/py_helpers/pyPandoraHelper/pyPandoraHelper.py"
 
-## This script removes the results for an individiaul while maintaining the nextflow process cache for them.
+## This script removes the results for an individual while maintaining the nextflow process cache for them.
 ##    It is intended as a way to refresh the results directories of an individual. This can be useful either
 ##    to remove older files after additional libraries appear and are therefore merged, or to remove results
 ##    with misleading names in cases where Pandora entries get updated (e.g. protocol mixup leading to changes
 ##    in strandedness for a library).
 
+## DEPENDENCY
+pandora_helper="/mnt/archgen/tools/helper_scripts/py_helpers/pyPandoraHelper/pyPandoraHelper.py"
+
+valid_analysis_types=("TF" "SG" "RP" "RM" "IM" "YC")
+
+## Join array elements by separator given as $1
+function join_array_elements() {
+  local IFS="$1"
+  shift
+  echo "$*"
+}
+
 ## Helptext function
 function Helptext() {
-  echo -ne "\t usage: $0 [options] <ind_id_list>\n\n"
-  echo -ne "This script removes all output directory contents for the provided individuals, without clearing out caching, allowing for the results to be re-published.\n    This enables refreshing of result directories when changes to the input might have changes merging of libraries, thus making the directory structure inconsistent.\n\n"
-  echo -ne "Options:\n"
-  echo -ne "-h, --help\t\tPrint this text and exit.\n"
-  echo -ne "-a, --analysis_type\t\tSet the analysis type. Options: TF, SG, RP, RM.\n"
+  errecho "\t usage: $0 [options] <ind_id_list>\n"
+  errecho "This script removes all output directory contents for the provided individuals, without clearing out caching, allowing for the results to be re-published.\n    This enables refreshing of result directories when changes to the input might have changes merging of libraries, thus making the directory structure inconsistent.\n"
+  errecho "Options:"
+  errecho "-h, --help\t\tPrint this text and exit."
+  errecho "-a, --analysis_type\t\tSet the analysis type. Options: $(join_array_elements , ${valid_analysis_types[@]})."
 }
 
 ## Print messages to stderr, optionally with colours
@@ -36,6 +46,8 @@ function errecho() {
   elif [[ ${1} == '-r' ]]; then
     colour="${Red}"
     shift 1
+  else
+    colour="${Normal}"
   fi
   echo -e ${colour}$*${Normal} 1>&2
 }
@@ -68,9 +80,11 @@ fi
 if [[ ${analysis_type} == '' ]]; then
   errecho "No --analysis_type was provided.\n"
   Helptext
-elif [[ ${analysis_type} != "SG" && ${analysis_type} != "TF" && ${analysis_type} != "RP" && ${analysis_type} != "RM" ]]; then
-  errecho "analysis_type must be SG, TF, RP, or RM. You provided: ${analysis_type}\n"
+  exit 2
+elif [[ ! " ${valid_analysis_types[*]} " =~ " ${analysis_type} " ]]; then
+  errecho "analysis_type must be one of: $(join_array_elements , ${valid_analysis_types[@]}). You provided: ${analysis_type}\n"
   Helptext
+  exit 2
 fi
 
 root_eager_dir='/mnt/archgen/Autorun_eager/eager_outputs' ## Directory should include subdirectories for each analysis type (TF/SG) and sub-subdirectories for each site and individual.