+++ scripts/run_Eager.sh
@@ -4,6 +4,7 @@
 pandora_helper="/mnt/archgen/tools/helper_scripts/py_helpers/pyPandoraHelper/pyPandoraHelper.py"
 
 ## Defaults
+valid_analysis_types=("TF" "SG" "RP" "RM" "IM" "YC")
 rush=''
 array=''
 temp_file=''
@@ -38,7 +39,7 @@ Yellow=$(tput sgr0)'\033[1;33m' ## Yellow normal face
 
 ## Since I'm running through all data every time, the runtime of the script will increase marginally over time. 
 ## Maybe create a list of eager inputs that are newer than the MQC reports and use that to loop over?
-for analysis_type in "SG" "TF" "RP" "RM" "YC"; do
+for analysis_type in ${valid_analysis_types[@]}; do
     # echo ${analysis_type}
     analysis_profiles="${nextflow_profiles},${analysis_type}"
     # echo "${root_input_dir}/${analysis_type}"